+++ mobile/android/android-components/components/feature/session/src/test/java/mozilla/components/feature/session/behavior/EngineViewBottomBehaviorTest.kt
@@ -62,6 +62,8 @@ class FakeEngineView(context: Context) : TextView(context), EngineView {
     override fun captureThumbnail(onFinish: (Bitmap?) -> Unit) {}
 
     override fun setVerticalClipping(clippingHeight: Int) {}
+
+    override fun release() {}
 }
 
 class BrowserToolbar(context: Context) : TextView(context)